requirements.txt
bs4
fastapi
uvicorn
requests
pydantic
gunicorn
python-dotenv
aiohttp